update employeedetails, fix product filter

+++ Project POS/POS/POS/EmployeeWorkSpace/UcMenu.xaml.cs
@@ -29,12 +29,12 @@ namespace POS.EmployeeWorkSpace
             try
             {
                 _unitofwork = ((MainWindow)Window.GetWindow(this))._unitofwork;
-                lvCategoryBeverages.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Drink);
-                lvCategoryDishes.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Food);
-                lvCategoryBeer.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Beer);
-                lvCategoryWine.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Wine);
-                lvCategorySnack.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Snack);
-                lvCategoryOther.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Other);
+                lvCategoryBeverages.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Drink);
+                lvCategoryDishes.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Food);
+                lvCategoryBeer.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Beer);
+                lvCategoryWine.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Wine);
+                lvCategorySnack.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Snack);
+                lvCategoryOther.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Other);
 
                 if (((MainWindow) Window.GetWindow(this)).currentTable == null)
                 {
@@ -152,12 +152,12 @@ namespace POS.EmployeeWorkSpace
 
             if(filter.Length == 0)
             {
-                lvCategoryBeverages.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Drink);
-                lvCategoryDishes.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Food);
-                lvCategoryBeer.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Beer);
-                lvCategoryWine.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Wine);
-                lvCategorySnack.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Snack);
-                lvCategoryOther.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Other);
+                lvCategoryBeverages.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Drink);
+                lvCategoryDishes.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Food);
+                lvCategoryBeer.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Beer);
+                lvCategoryWine.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Wine);
+                lvCategorySnack.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Snack);
+                lvCategoryOther.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Other);
                 return;
             }
 
@@ -170,42 +170,42 @@ namespace POS.EmployeeWorkSpace
         {
             if (ItemBeverages.IsSelected == true)
             {
-                lvCategoryBeverages.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Drink && p.Name.Contains(filter));
+                lvCategoryBeverages.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Drink && p.Name.Contains(filter));
                 lvCategoryBeverages.PreviewMouseLeftButtonUp += lvCategory_PreviewMouseLeftButtonUp;
                 curItem = ItemBeverages;
             }
 
             if (ItemDishes.IsSelected == true)
             {
-                lvCategoryDishes.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Food && p.Name.Contains(filter));
+                lvCategoryDishes.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Food && p.Name.Contains(filter));
                 lvCategoryDishes.PreviewMouseLeftButtonUp += lvCategory_PreviewMouseLeftButtonUp;
                 curItem = ItemDishes;
             }
 
             if (ItemBeer.IsSelected == true)
             {
-                lvCategoryBeer.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Beer && p.Name.Contains(filter));
+                lvCategoryBeer.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Beer && p.Name.Contains(filter));
                 lvCategoryBeer.PreviewMouseLeftButtonUp += lvCategory_PreviewMouseLeftButtonUp;
                 curItem = ItemBeer;
             }
 
             if (ItemWine.IsSelected == true)
             {
-                lvCategoryWine.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Wine && p.Name.Contains(filter));
+                lvCategoryWine.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Wine && p.Name.Contains(filter));
                 lvCategoryWine.PreviewMouseLeftButtonUp += lvCategory_PreviewMouseLeftButtonUp;
                 curItem = ItemWine;
             }
 
             if (ItemSnack.IsSelected == true)
             {
-                lvCategorySnack.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Snack && p.Name.Contains(filter));
+                lvCategorySnack.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Snack && p.Name.Contains(filter));
                 lvCategoryWine.PreviewMouseLeftButtonUp += lvCategory_PreviewMouseLeftButtonUp;
                 curItem = ItemSnack;
             }
 
             if (ItemOther.IsSelected == true)
             {
-                lvCategoryOther.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == ProductType.Other && p.Name.Contains(filter));
+                lvCategoryOther.ItemsSource = _unitofwork.ProductRepository.Get(p => p.Type == (int)ProductType.Other && p.Name.Contains(filter));
                 lvCategoryOther.PreviewMouseLeftButtonUp += lvCategory_PreviewMouseLeftButtonUp;
                 curItem = ItemOther;
             }
